Improve DAS5 sync script

+++ utils/sync_das5.sh
@@ -9,9 +9,15 @@ if [ "$(basename "$PWD")" != "DLSA" ]; then
     exit 1
 fi
 
-cd rust
-make build-linux-avx2 # The docker daemon should be running for this.
-cd ..
-rsync -ru --delete --progress --exclude='/.git' --filter="dir-merge,- .gitignore" ./* DAS5:DLSA
+if ! docker info &> /dev/null; then
+  echo "Docker daemon is not running. Exiting..."
+  exit 1
+fi
+
+cd rust &&
+make build-linux-avx2 && # The docker daemon should be running for this.
+cd .. &&
+rsync -ru --delete --progress --exclude='/.git' --filter="dir-merge,- .gitignore" ./* DAS5:DLSA &&
 
 rsync -ru --delete --progress --exclude '*/' rust/target/x86_64-unknown-linux-gnu/release/* DAS5:DLSA/rust/target/release
+# Make sure the directories used here exist (on the host and on the DAS5 remote).
